backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
python-multipart>=0.0.9
google-generativeai>=0.5.0